Incomplete but substantial first pass at CI/infra
This commit introduces the first CI code and infrastructure code into
the project. CI is doing all the typical stuff: linting, testing,
building javascript. But we've also defined the infrastructure for the
project via terraform and are linting, planning, and applying that.

Finally we have CI steps that use aws orbs to put our applications on
the infra. The frontend app on s3 and an image of the server in ecr. The
only remaining step is to define an ecs cluster in terraform and have
circle take the server image from ecr and put it on ecs.